Develop

+++ Diego-Lofiego/Modelo/Usuario.cs
@@ -2,14 +2,14 @@
 {
     public class Usuario : IId
     {
-        private int _id;
+        private long _id;
         private string _nombre;
         private string _apellido;
         private string _nombreUsuario;
         private string _contraseña;
         private string _mail;
 
-        public int Id { get { return _id; } set { _id = value; } }
+        public long Id { get { return _id; } set { _id = value; } }
         public string Nombre { get { return _nombre; } set { _nombre = value; } }
         public string Apellido { get { return _apellido; } set { _apellido = value; } }
         public string NombreUsuario { get { return _nombreUsuario; } set { _nombreUsuario = value; } }
@@ -26,7 +26,7 @@
             _mail = String.Empty;
         }
 
-        public Usuario(int id, string nombre, string apellido, string nombreUsuario, string contraseña, string mail)
+        public Usuario(long id, string nombre, string apellido, string nombreUsuario, string contraseña, string mail)
         {
             _id = id;
             _nombre = nombre;
